code/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
requests
openai
firebase-admin
python-dotenv
aiofiles
pydub
pytest
pytest-asyncio
httpx
matplotlib
groq
jiwer
librosa
numpy
pandas
pydub
seaborn
soundfile
